typings + typedoc for AreaShape/LinkShape/LineShape
